Refactor loading of posts into reusable function
